move stats package to logqlmodel
