+++ pkg/logqlmodel/logqlmodel.go
@@ -4,7 +4,7 @@ import (
 	"github.com/prometheus/prometheus/promql/parser"
 
 	"github.com/grafana/loki/pkg/logproto"
-	"github.com/grafana/loki/pkg/logql/stats"
+	"github.com/grafana/loki/pkg/logqlmodel/stats"
 )
 
 // ValueTypeStreams promql.ValueType for log streams
