+++ pkg/logqlmodel/stats/stats.proto
